FIX: Flag variable overflow

+++ gui/halt_list_frame.h
@@ -83,7 +83,7 @@ private:
 	static sort_mode_t sortby;
 	static bool sortreverse;
 
-	static uint8 filter_flags;
+	static uint32 filter_flags;
 
 	static char name_filter_value[64];
